+++ src/apps/badger/badger.definition.ts
@@ -25,7 +25,10 @@ export const BADGER_DEFINITION = appDefinition({
 
   tags: [AppTag.YIELD_AGGREGATOR],
   keywords: [],
-  links: {},
+  links: {
+    twitter: 'https://twitter.com/badgerdao',
+    discord: 'https://discord.com/invite/xSPFHHS',
+  },
 
   supportedNetworks: {
     [Network.ETHEREUM_MAINNET]: [AppAction.VIEW],