+++ src/apps/stargate/arbitrum/stargate.pool.token-fetcher.ts
@@ -1,11 +1,7 @@
-import { Inject } from '@nestjs/common';
-
 import { Register } from '~app-toolkit/decorators';
-import { PositionFetcher } from '~position/position-fetcher.interface';
-import { AppTokenPosition } from '~position/position.interface';
 import { Network } from '~types/network.interface';
 
-import { StargatePoolTokenHelper } from '../helpers/stargate.pool.token-helper';
+import { StargatePoolTokenFetcher } from '../common/stargate.pool.token-fetcher';
 import { STARGATE_DEFINITION } from '../stargate.definition';
 
 const appId = STARGATE_DEFINITION.id;
@@ -13,13 +9,10 @@ const groupId = STARGATE_DEFINITION.groups.pool.id;
 const network = Network.ARBITRUM_MAINNET;
 
 @Register.TokenPositionFetcher({ appId, groupId, network })
-export class ArbitrumStargatePoolTokenFetcher implements PositionFetcher<AppTokenPosition> {
-  constructor(@Inject(StargatePoolTokenHelper) private readonly stargatePoolTokenHelper: StargatePoolTokenHelper) {}
-
-  async getPositions() {
-    return this.stargatePoolTokenHelper.getPositions({
-      factoryAddress: '0x55bdb4164d28fbaf0898e0ef14a589ac09ac9970',
-      network,
-    });
-  }
+export class ArbitrumStargatePoolTokenFetcher extends StargatePoolTokenFetcher {
+  appId = appId;
+  groupId = groupId;
+  network = network;
+  factoryAddress = '0x55bdb4164d28fbaf0898e0ef14a589ac09ac9970';
+  useLocalDecimals = false;
 }