+++ src/apps/synthetix/optimism/synthetix.synth.token-fetcher.ts
@@ -20,6 +20,7 @@ export class OptimismSynthetixSynthTokenFetcher implements PositionFetcher<AppTo
     return this.tokenHelper.getTokens({
       network,
       resolverAddress: '0x95a6a3f44a70172e7d50a9e28c85dfd712756b8c',
+      exchangeable: true,
     });
   }
 }