+++ src/apps/yearn/yearn.module.ts
@@ -4,7 +4,6 @@ import { SynthetixAppModule } from '~apps/synthetix';
 
 import { ArbitrumYearnV2VaultTokenFetcher } from './arbitrum/yearn.v2-vault.token-fetcher';
 import { YearnContractFactory } from './contracts';
-import { EthereumYearnBalanceFetcher } from './ethereum/yearn.balance-fetcher';
 import { EthereumYearnGovernanceContractPositionFetcher } from './ethereum/yearn.governance.contract-position-fetcher';
 import { EthereumYearnV1VaultTokenFetcher } from './ethereum/yearn.v1-vault.token-fetcher';
 import { EthereumYearnV2VaultTokenFetcher } from './ethereum/yearn.v2-vault.token-fetcher';
@@ -22,7 +21,6 @@ import { YearnAppDefinition, YEARN_DEFINITION } from './yearn.definition';
     // Helpers
     YearnVaultTokenDefinitionsResolver,
     // Ethereum
-    EthereumYearnBalanceFetcher,
     EthereumYearnV1VaultTokenFetcher,
     EthereumYearnV2VaultTokenFetcher,
     EthereumYearnGovernanceContractPositionFetcher,